# openapi/messages/openapi.yaml
openapi: 3.0.0
info:
  title: Messages API
  version: 1.0.0
  description: API для управления сообщениями

paths:
  /messages:
    get:
      summary: Get all messages
      tags:
        - messages
      responses:
        "200":
          description: A list of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      summary: Create a new message
      tags:
        - messages
      requestBody:
        description: The message to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Message"
      responses:
        "201":
          description: The created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /messages/{id}:
    patch:
      summary: Update an existing message
      tags:
        - messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        description: The updated message data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Message"
      responses:
        "200":
          description: The updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      summary: Delete a message
      tags:
        - messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "204":
          description: Message successfully deleted

components:
  schemas:
    Message:
      type: object
      properties:
        id:
          type: integer
          format: uint
        message:
          type: string
      required:
        - message